UI Overhaul
Decided what I want this program to do exactly. It just pulls sprites
from the file, and outputs to sheets of a specified size.

I've changed the UI to reflect this, and almost all of the functionality
is there. I have a bit more work to do on the math of how to place
sprites on the bitmap sheets, but this project is already 80% done.